+++ docs/Whats-New.md
@@ -605,7 +605,7 @@ HideShakeEffects=false           ; boolean
 - [Auto-remove earliest beacon](Fixed-or-Improved-Logics.md#auto-remove-earliest-beacon) (by TaranDahl)
 - [Customize the step limit of the credits indicator](User-Interface.md#customize-the-step-limit-of-the-credits-indicator) (by Noble_Fish)
 - [Disable the credits indicator smooth transition](User-Interface.md#disable-the-credits-indicator-smooth-transition) (by Noble_Fish)
-- [New passengers display feature](User-Interface.md#passengers-display) (by ABCZ)
+- [New passengers cameo display feature](User-Interface.md#passengers-display) (by ABCZ)
 
 #### Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)